// server/src/models/models.ts
export interface Recipe {
  id: string;
  imageUrl?: string;
  title: string;
  description?: string;
  ingredients: Ingredient[];
  instructions?: string[];
  updatedAt: Date;
}

export interface Ingredient {
  product: string;
  amount: number;
  unit: Unit;
}

export enum Unit {
  Gram = 'грамм',
  Kg = 'килограмм',
  Ml = 'мл',
  L = 'л.',
  Piece = 'штука',
  Tsp = 'ч.л.',
  Tbsp = 'ст.л.',
  Glass = 'стакан',
}